fix: address PR #149 review — remove noqa, fix types, add serve tests, enforce line limit
